fix some warnings in VS2019

+++ src/core/accelerator.h
@@ -39,11 +39,11 @@ constexpr int ACCELERATOR_SAH_BUCKET_NUM = 12;
 
 struct BVHPrimitiveInfo
 {
-    size_t prim_index;
+	uint32_t prim_index;
     Bounds3f bounds;
     Point3f centroid;
     BVHPrimitiveInfo() = default;
-    BVHPrimitiveInfo(const Primitive &p, size_t index) : prim_index(index), bounds(get_world_bounds(p)), centroid((get_world_bounds(p).min_point + get_world_bounds(p).max_point) * 0.5f) {}
+    BVHPrimitiveInfo(const Primitive &p, uint32_t index) : prim_index(index), bounds(get_world_bounds(p)), centroid((get_world_bounds(p).min_point + get_world_bounds(p).max_point) * 0.5f) {}
 };
 
 struct BVHBuildNode
@@ -221,7 +221,7 @@ private:
     std::vector<SoAPrimitiveInfo> _soa_primitive_infos;
     std::vector<QBVHNode> _nodes;
 
-    BVHBuildNode *build(MemoryArena &arena, size_t start, size_t end, std::vector<BVHPrimitiveInfo> &primitive_infos, std::vector<Primitive> &ordered, uint32_t *total,ProgressReporter* reporter);
+    BVHBuildNode *build(MemoryArena &arena, uint32_t start, uint32_t end, std::vector<BVHPrimitiveInfo> &primitive_infos, std::vector<Primitive> &ordered, uint32_t *total,ProgressReporter* reporter);
     void build_soa_primitive_info(BVHBuildNode *node);
     QBVHCollapseNode *collapse(MemoryArena &arena, const BVHBuildNode *subtree_root, uint32_t *total);
     uint32_t flatten(const QBVHCollapseNode *c_node, uint32_t *offset);
